Add sorting functionality
User might want to easily see the most or least urgent task to do.

A sort command is added to duke. Users can access this command by typing
"sort ascending" or "sort descending" to order the tasks in the list
from latest to earliest and vice versa.

Giving user the option to choose ascending and descending allow for more
flexibility and convenience especially when the list is long.